xfers_recv: reject NULL progress cb; cast 32-bit HTXF preamble size
Review fixes:

- file_recv_one called progress unconditionally; a NULL hook would
  segfault. Reject it up front with EINVAL and document the contract in
  the header.

- test_file_get passed a guint64 to integration_send_xfer_hdr, whose
  legacy 16-byte HTXF preamble carries a 32-bit size — a narrowing
  conversion (warning under -Werror). Cast explicitly; the size is
  already capped < 1 MiB.

+++ tests/integration/test_file_get.c
@@ -138,7 +138,10 @@ test_file_get_round_trip (void)
         integration_close (fd);
         return;
     }
-    g_assert_true (integration_send_xfer_hdr (xfd, xfer_ref, xfer_size));
+    /* The legacy 16-byte HTXF preamble carries a 32-bit size; xfer_size
+     * is capped < 1 MiB above, so the narrowing cast is safe. */
+    g_assert_true (
+        integration_send_xfer_hdr (xfd, xfer_ref, (guint32)xfer_size));
 
     /* Wrap the connected fd as an hxnet HTXF channel (plaintext, no
      * preamble — we already sent it above). hxnet adopts xfd; the
